WritingAssistantBackend/pyproject.toml
[project]
name = "writingassistantbackend"
version = "0.1.0"
description = ""
authors = [
    {name = "Philippe-Lauwers",email = "philippe.lauwers@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "numpy (>=1.25.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "kenlm (==0.3.0)",
    "opennmt-py (==1.1.1)",
    "torch @ https://download.pytorch.org/whl/cu118/torch-2.6.0%2Bcu118-cp310-cp310-win_amd64.whl",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
